+++ view/packages/wikibase-data-values-value-view/src/experts/MonolingualText.js
@@ -46,11 +46,9 @@
 		 * @inheritdoc
 		 */
 		valueCharacteristics: function() {
-			var options = {};
-			if ( this._languageSelector ) {
-				options.valuelang = this._languageSelector.getValue();
-			}
-			return options;
+			return {
+				valuelang: this._languageSelector.getValue()
+			};
 		},
 
 		/**